# src/api-spec.yaml
openapi: 3.0.0
info:
  title: EffesEYE Platform API
  version: 2020-10
  contact:
    email: chaluwa@gmail.com
    name: Charles Odili
  license:
    name: MIT
  description: |-
    Charles Odili just joined a fintech startup, EffesEYE, and was asked to lead the development of an urgent project on the table, a simple billing application.

    In order to have a view on the practicality of this new product, the Board asked that he develop and present an MVP of this product to them at the next Management meeting.

    Odili was required to build an MVP on the [FSI Sandbox](https://sandbox.fsi.ng/), integrated into the APIs of the available Banks and services of the Sandbox. This application can buy telephone credit, pay electricity bills and Pay TV bills (DSTV, Startimes, etc).


    Odili adopted an API-first approach, and this is the documentation convering the API that drives the services and features the EffesEYE billing (MVP) platform provides to customers.

    The API is deployed at [https://effeseye-api.herokuapp.com](https://effeseye-api.herokuapp.com), is managed by a secure admin application at [https://effeseye-admin.netlify.app](https://effeseye-admin.netlify.app), and is available for customers via an installable Progressive Web App at [https://effeseye-admin.netlify.app](https://effeseye-admin.netlify.app) 
  termsOfService: 'https://effeseye-api.herokuapp.com/tc'
servers:
  - url: 'https://effeseye-api.herokuapp.com'
    description: Production
paths:
  '/{version}/pay/{service}':
    post:
      summary: Payment service endpoint
      tags:
        - airtime
        - electricity
        - paytv
        - payments
      responses:
        '201':
          description: Created. Raised when the payment completed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaymentResponse'
        '401':
          description: Unauthorized. Usually raised due to supplying an invalid access token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      operationId: post-airtime-recharge
      description: 'Buy airtime credit, pay for electricity or TV subscription with this endpoint'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PaymentRequest'
            examples:
              airtime-example:
                value:
                  amount: 1000
                  currency: NGN
                  provider: Airtel
                  sendTo: '+2348138145472'
                  accountToDebit: '0037514056'
              electricity-example:
                value:
                  amount: 8500
                  currency: NGN
                  provider: EKDC
                  sendTo: 56834930-5-4564-8-gbfg
                  accountToDebit: '0037514056'
              tv-example:
                value:
                  amount: 12000
                  currency: NGN
                  provider: DSTV
                  sendTo: 45783-08-fjsd0948503
                  accountToDebit: '0037514056'
        description: HTTP request payload for the payment service
      security:
        - Authorization: []
      parameters: []
    parameters:
      - schema:
          type: string
          minLength: 7
          maxLength: 7
          pattern: '^\d{4}-\d{2}$'
          example: 2020-10
        name: version
        in: path
        required: true
        description: The API version
      - schema:
          type: string
          enum:
            - airtime
            - electricity
            - tv
          minLength: 2
          maxLength: 11
        name: service
        in: path
        required: true
        description: The service you want to pay for
  '/{version}/pay/{paymentId}/status':
    parameters:
      - schema:
          type: string
        name: paymentId
        in: path
        required: true
        description: The request id of a previous airtime recharge request
      - schema:
          type: string
          minLength: 7
          maxLength: 7
          pattern: '^\d{4}-\d{2}$'
          example: 2020-10
        name: version
        in: path
        required: true
        description: The API version
    get:
      summary: Payment enquiry / status endpoint
      tags:
        - airtime
        - electricity
        - paytv
        - payments
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaymentResponse'
              examples:
                example:
                  value:
                    message: payment status
                    paymentId: 45934508043-583
                    status: SUCCEEDED
        '401':
          description: Unauthorized. Usually raised due to supplying an invalid access token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      operationId: get-airtime-recharge-requestId
      description: Query the status of a payment transaction
      security:
        - Authorization: []
      parameters: []
...
  '/{version}/bo/users':
    parameters:
      - schema:
          type: string
          example: 2020-10
          pattern: '^\d{4}-\d{2}$'
          minLength: 7
          maxLength: 7
        name: version
        in: path
        required: true
        description: The API version
    get:
      summary: Fetch user account data
      tags:
        - admin
        - user-data
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UsersListingResponse'
              examples:
                example:
                  value:
                    message: success
                    data:
                      - accountid: 45930953573nvvgjout4309u045043
                        firstname: Charles
                        lastname: Odili
                        accounttype: ADMIN
                        email: chalu@example.com
                        lastseen: '2020-10-24T15:38:12.580Z'
                        membersince: '2020-10-24T15:38:12.580Z'
        '401':
          description: Unauthorized. Usually raised due to supplying an invalid access token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      operationId: get-version-bo-users
      description: Fetch a collection of registered users (accounts) on the platform
      parameters: []
      security:
        - Authorization: []
...
  '/{version}/account/add-bank':
    post:
      summary: Verify and add a (NUBAN) bank account
      tags:
        - NUBAN
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AddBankAccountResponse'
        '401':
          description: Unauthorized. Usually raised due to supplying an invalid access token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: 'Forbidden. Raised if the NUBAN account number cannot be verifed, or the related EffesEYE account details are invalid'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      operationId: post-account-add-bank
      security:
        - Authorization: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddBankAccountRequest'
            examples:
              example:
                value:
                  nuban: '0037514056'
                  bankname: UBA
        description: HTTP request payload to add a bank account
      description: 'Verify and add a NUBAN (bank account) number to a user''s EffesEYE account, allowing them initiate payments from the bank accounts'
    parameters:
      - schema:
          type: string
          pattern: '^\d{4}-\d{2}$'
          minLength: 7
          maxLength: 7
          example: 2020-10
        name: version
        in: path
        required: true
        description: The API version
components:
  schemas:
    PaymentResponse:
      title: PaymentResponse
      type: object
      description: HTTP response payload for making a payment
      x-tags:
        - airtime
      x-examples: {}
      properties:
        message:
          type: string
          description: a message describing the status or progress of the payment
        paymentId:
          type: string
          description: The id of the payment request
        status:
          type: string
          description: 'The status of the payment request. Can be FAILED, PENDING, or SUCCEEDED'
          enum:
            - PENDING
            - FAILED
            - SUCCEEDED
        details:
          type: object
          description: An optional object containing additional information partaining to the payment attempted
      required:
        - message
        - paymentId
        - status
    PaymentRequest:
      title: PaymentRequest
      type: object
      x-tags:
        - airtime
      description: The HTTP request payload for making a payment
      properties:
        amount:
          type: integer
          description: The amount to pay for the service
        currency:
          type: string
          enum:
            - NGN
            - USD
          description: The ISO 4217 currency code the `amount` is expressed in
          minLength: 3
          maxLength: 3
        provider:
          type: string
          description: |-
            The provider whose service is being purchased:  

            > For airtime, this could be `Airtel`, `MTN`, `Glo`, `9Mobile` e.t.c.

            > For PayTV, this could be `DSTV`, `Startimes`, `GoTV` e.t.c

            > For electricity, this could be `EKDC`, `IKEDC`, `BEDC`, `IBEDC` e.t.c
        sendTo:
          type: string
          description: 'The corresponding phone number, meter number, or TV subscription Id to credit this recharge to'
        accountToDebit:
          type: string
          description: The verified (NUBAN) bank account to debit for this payment
          minLength: 10
          maxLength: 10
      required:
        - amount
        - currency
        - provider
        - sendTo
        - accountToDebit
    ErrorResponse:
      title: ErrorResponse
      type: object
      description: The HTTP response payload for requests that terminated with errors
      properties:
        message:
          type: string
          description: A statement or description of the error
      required:
        - message
...
    UsersListingResponse:
      title: UsersListingResponse
      type: object
      description: HTTP response payload for fetching a list of user (account) data
      x-tags:
        - admin
        - user-data
      properties:
        message:
          type: string
          description: A message describing the handling of the request. Will be helful if there was an error
        data:
          type: array
          description: An array of user account data objects
          items:
            $ref: '#/components/schemas/UserData'
      required:
        - message
    UserData:
      title: UserData
      type: object
      description: Data represneting a user in the platform
      x-tags:
        - admin
        - user-data
      properties:
        accountid:
          type: string
          description: The account-id of the user
        firstname:
          type: string
          description: The user's firstname
        lastname:
          type: string
          description: The user's lastname
        email:
          type: string
          description: The user's email address
        lastseen:
          description: The timestamp of when the user last logged into the platform
          type: string
        membersince:
          description: The timestamp of when the account was created
          type: string
        accounttype:
          type: string
          enum:
            - USER
            - ADMIN
          description: The user account type
      required:
        - accountid
        - firstname
        - lastname
        - email
        - lastseen
        - membersince
        - accounttype
...
    AddBankAccountRequest:
      title: AddBankAccountRequest
      type: object
      description: HTTP request payload to add a bank account
      properties:
        nuban:
          type: string
          minLength: 10
          maxLength: 10
          description: The NUBAN account number
        bankname:
          type: string
          minLength: 3
          description: The name of the Bank
      required:
        - nuban
        - bankname
    AddBankAccountResponse:
      title: AddBankAccountResponse
      type: object
      description: HTTP response payload from attempting to verify and add a NUBAN account
      properties:
        message:
          type: string
          description: description of handling the request
        status:
          type: string
          enum:
            - FAILED
            - SUCCEEDED
          description: The status of the operation
      required:
        - message
        - status
  securitySchemes:
    Authorization:
      type: http
      scheme: bearer
      description: The HTTP Authorization header containing a valid access token designated as `Bearer`
...